✨ FEAT: 업무유형 리드테이블 컴포넌트 생성
