+++ src/components/feature/table/taskDetailTable.css
@@ -1,0 +1,3 @@
+.tableP {
+  margin: 0px 0px;
+}
